# swagger.yaml
swagger: "2.0"
info:
  description: "The REST API of Production Manager"
  version: "1.0.0"
  title: "ProductionManager"
basePath: "/api/"
...
schemes:
- "http"
paths:
...
  /orders/{orderId}:
     get:
       tags:
       - "orders"
       summary: "get all orders"
       parameters:
         - name: "orderId"
           in: "path"
           description: "ID of order"
           required: true
           type: "string"
       produces:
         - application/json
       responses:
         200:
           description: "all order"
           schema:
             $ref: '#/definitions/Order'
     delete:
        tags:
        - "orders"
        summary: "delete all orders"
        parameters:
         - name: "orderId"
           in: "path"
           description: "ID of order"
           required: true
           type: "string"
        produces:
          - application/json
        responses:
         200:
           description: "order deleted"
  /orders/{orderId}/licenseupdate:
     get:
       tags:
       - "orders"
       summary: "trigger licenseupdate"
       parameters:
         - name: "orderId"
           in: "path"
           description: "ID of order"
           required: true
           type: "string"
       produces:
         - application/json
       responses:
         200:
           description: "triggered licenseupdate"
...
definitions:
...
  Order:
    type: object
    properties:
      items:
        type: array
        items:
          type: object
          properties:
            objectId:
              type: string
            amount:
              type: integer
      offer:
        type: object
        properties:
          id: 
            type: string
          bip21:
            type: string
      state:
        type: string
      createdAt:
        type: string
...